Add complete modal and fix the way scoring is performed

+++ server/common/models/UserStoryAttempt.py
@@ -12,7 +12,7 @@ class UserStoryAttempt(db.Model):
     short_story_id = db.Column(db.Integer(), db.ForeignKey('short_story.id'))
     short_story_translation_id = db.Column(db.Integer(), db.ForeignKey('short_story_translation.id'))
     language = db.Column(db.String(10))
-    score = db.Column(db.Integer())
+    score = db.Column(db.Float())
     responses = db.Column(db.JSON())
     is_complete = db.Column(db.Boolean())
     created_at = db.Column(db.DateTime(timezone=True), server_default=func.now())
